requirements.txt
Django>=4.2
djangorestframework
channels
channels-redis
daphne
psycopg2-binary
Pillow
django-cors-headers